+++ Changelog.md
@@ -12,6 +12,7 @@
 - split large oak trees into two nbt's
     - adjusts better to terrain adaptions
 - add statues that can generate in villages
+- add new structure campsite
 
 ### Version 1.19.2 - 1.5.2